Add block access

+++ src/filesys/hfs/fileadaptor.rs
@@ -6,7 +6,7 @@ pub trait FileAccess : std::fmt::Debug {
     fn seek(&mut self, pos : u64) -> io::Result<u64>;
     fn size(&mut self) -> io::Result<u64>;
     fn pos(&mut self) -> io::Result<u64>;
-    fn read(&mut self, len : usize) -> io::Result<Vec<u8>>;
+    fn read(&mut self, len : u64) -> io::Result<Vec<u8>>;
 }
 
 impl<'storage, T> FileAdaptor<'storage, T>
@@ -35,8 +35,8 @@ T: io::Read + io::Seek {
     fn pos(&mut self) -> io::Result<u64> {
         self.0.seek(io::SeekFrom::Current(0))
     }
-    fn read(&mut self, len : usize) -> io::Result<Vec<u8>> {
-        let mut bufv : Vec<u8> = Vec::with_capacity(len);
+    fn read(&mut self, len : u64) -> io::Result<Vec<u8>> {
+        let mut bufv : Vec<u8> = Vec::with_capacity(len as usize);
         for _ in 0..len {
             let mut arr = [0u8;1];
             self.0.read_exact(&mut arr)?;
